m12n: Move PerfLogger to oak-commons
