+++ oak-commons/src/main/java/org/apache/jackrabbit/oak/commons/benchmark/PerfLogger.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.jackrabbit.oak.util;
+package org.apache.jackrabbit.oak.commons.benchmark;
 
 import java.util.concurrent.TimeUnit;
